+++ src/parser/block/util/index.test.ts
@@ -12,6 +12,7 @@ describe('Unit: parser/block/util/indexer', () => {
       assert.deepStrictEqual(inspect(parser(' ')), undefined);
       assert.deepStrictEqual(inspect(parser(' #')), undefined);
       assert.deepStrictEqual(inspect(parser(' #a')), undefined);
+      assert.deepStrictEqual(inspect(parser(' [#a]]')), undefined);
     });
 
     it('ab', () => {