+++ src/parser/inline/extension/index.ts
@@ -1,18 +1,18 @@
 ﻿import { ExtensionParser } from '../../inline';
-import { loop } from '../../../combinator';
 import { link } from '../link';
-import { text } from '../../source/text';
-import { squash } from '../../squash';
 import { isTightVisible } from '../../source/validation';
-import { makeIndex } from '../../string/index';
+import { defineIndex } from '../../block/util/index';
 import { template } from './template';
 
-export const index: ExtensionParser.IndexParser = template((flag: string, query: string): [[HTMLAnchorElement], string] | undefined => {
+export const index: ExtensionParser.IndexParser = template((flag, query, frag): [[HTMLAnchorElement], string] | undefined => {
   if (flag !== '#') return;
-  const [[el = undefined] = [], rest = ''] = link(`[](#)`) || [];
+  if (frag.textContent === flag) return;
+  const [[el = undefined] = [], rest = ''] = link(`[${query}]()`) || [];
   if (!el) return;
-  el.textContent = squash((loop(text)(query) || [[]])[0]).textContent!;
-  if (!isTightVisible(el.textContent)) return;
-  void el.setAttribute('href', `#${makeIndex(el.textContent)}`);
+  assert(rest === '');
+  if (!isTightVisible(el.textContent!)) return;
+  void defineIndex(el);
+  void el.setAttribute('href', `#${el.id}`);
+  void el.removeAttribute('id');
   return [[el], rest];
 });